refactor: create `downloadEntryCover` helper function
